+++ src/main/kotlin/me/gurkz/superdupermod/SuperDuperMod.kt
@@ -12,6 +12,9 @@ import me.gurkz.superdupermod.command.OfflineTpCommand
 import me.gurkz.superdupermod.command.ServerSayCommand
 import me.gurkz.superdupermod.command.SmiteCommand
 import me.gurkz.superdupermod.command.SuicideCommand
+import me.gurkz.superdupermod.data.DataStorages
+import me.gurkz.superdupermod.data.DeathLocationData
+import me.gurkz.superdupermod.event.DeathLocationEventListeners
 import me.gurkz.superdupermod.event.OfflineTpEventListeners
 import me.gurkz.superdupermod.item.ModItems
 import me.gurkz.superdupermod.network.RespawnPlayer
@@ -40,6 +43,9 @@ object SuperDuperMod : ModInitializer {
         SuicideCommand.register()
         OfflineTpCommand.register()
         OfflineTpEventListeners.register()
+
+        DataStorages.register()
+        DeathLocationEventListeners.register()
 
         ModItems.initialise()